requirements.txt
tqdm
torch
torchvision
opencv-python
pillow
numpy
gradio